Renamed ARCMacros.h
